# eakte-api/work-in-progress/api-doc 2026-08-28.yaml
openapi: 3.0.0
info:
  title: DMS REST API
  version: 1.2.8-snapshot
servers:
  - url: /api
...
paths:
...
  '/setSignatureObjects/{objaddress}':
    put:
      tags:
        - setSignatureObjects
      summary: Unterschrift setzen auf Objekten
      operationId: SetSignature
      parameters:
        - name: x-anwendung
          in: header
          description: Aufrufende Fachanwendung
          required: false
          schema:
            type: string
          example: SWH
        - name: userlogin
          in: header
          description: >-
            Übergabe des Benutzer-Logins, von deren Berechtigung die Aktion
            durchgeführt werden soll.
          required: false
          schema:
            type: string
          example: max.mustermann
        - name: joboe
          in: header
          description: COO-ID der Organisationseinheit des Nutzers
          required: false
          schema:
            type: string
          example: COO.1.2300.1.847
        - name: jobposition
          in: header
          description: 'Rolle, in der der Nutzer handelt'
          required: false
          schema:
            type: string
          example: DocumentManager
        - name: objaddress
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSignatureRequestDTO'
        required: true
      responses:
        '204':
          description: success
        '400':
          description: Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmsErrorResponse'
        '401':
          description: Falscher oder fehlender technischer Nutzer.
        '500':
          description: Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten.
      security:
        - basicAuthSecurityDefintion: []
...
components:
  schemas:
...
    DmsErrorResponse:
      type: object
      properties:
        fehlerQuelle:
          type: string
          enum:
            - INTERN
            - DMS
        status:
          type: integer
          format: int32
        text:
          type: string
...
    SetSignatureRequestDTO:
      required:
        - signaturetype
      type: object
      properties:
        signaturetype:
          type: string
          description: >-
            COO-Adresse der Unterschriftenart

            Anmerkung: Titel ist nicht eindeutig, da es im System mehrere
            Unterschriftenarten mit dem gleichen Namen/Titel gibt.
          example: COO.1.1001.1.6485
        remark:
          type: string
          description: Bemerkung zur Unterschrift
          example: String
...
  securitySchemes:
    basicAuthSecurityDefintion:
      type: http
      description: >-
        This API uses Basic Auth flow. [More
        info](https://api.example.com/docs/auth)
      scheme: basic