# v1/eakte-api-v1/eakte-api-v1-spring-boot-server-stubs/src-generated/src/main/java/de/muenchen/itm/eakte/schnittstelle/rest_v2/server_stubs/controllers/SetSignatureObjectsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.controllers;

import de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.model.DmsErrorResponse;
import org.springframework.lang.Nullable;
import de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.model.SetSignatureRequestDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;
import jakarta.servlet.http.HttpServletRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.21.0")
@Validated
@Controller
@Tag(name = "/setSignatureObjects", description = "the /setSignatureObjects API")
public interface SetSignatureObjectsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_SET_SIGNATURE = "/setSignatureObjects/{objaddress}";
    /**
     * PUT /setSignatureObjects/{objaddress} : Unterschrift setzen auf Objekten
     * Bei dem Schriftgutobjekt (parameter: objaddress) wird eine definierte Unterschriftenart gesetzt (parameter: signaturetype).  Dabei wird die in der Domäne/Mandanten gültige Unterschriftenregelung beachtet.  Die jeweiligen Vorbedingungen werden vor dem Anbringen der Unterschrift ebenfalls geprüft.  Bitte beachten Sie, dass beim Anbringen der Unterschrift  auch Folge-Aktivitäten ausgeführt werden  (wie z.B. der Bearbeitungsstatus wird korrekt gesetzt).\\ Unterschriften, in denen weitere GUI-Masken angezeigt werden (wie z.B. bei Umschreibungen),  können über diese Soap-Aktion natürlich nicht angebracht werden.  Für entsprechende Aktionen müssen bei Bedarf spezielle Soap-Aktionen entwickelt werden.  Im Kontext des 4-Augen-Löschens sind über diesen Aufruf  nur die dafür freigegebenen Unterschriften „Löschmarkierung setzen“, „Löschmarkierung bestätigen“  und „Löschmarkierung ablehnen“ über die Schnittstelle zulässig,  sofern sie in der gültigen Unterschriftenkonfiguration vorgesehen sind und  alle fachlichen Vorbedingungen erfüllt sind.\\ Die Aktivitäten im Arbeitsvorrat, die nach Löschmarkierung setzen gestartet werden,  werden im Arbeitsvorrat des Technischen Nutzers aufgeräumt,  wenn der AT-Job die bestätigten Löschaufträge aus dem Arbeitsvorrat entfernt. 
     *
     * @param userlogin **Benutzer-Login** (Benutzer-Kontext für den SOAP-Aufruf)\\ Log-in-Name des Benutzerobjekts  (required)
     * @param objaddress Objekt-ID (COO-Adresse) des Objekts (required)
     * @param setSignatureRequestDTO  (required)
     * @param xAnwendung **Aufrufende Fachanwendung**  Für jeden Aufruf soll ersichtlich sein,  welche Fachanwendung den Aufruf getätigt hat.  Dies soll es ermöglichen, Requests und Responses den entsprechenden Fachverfahren zuzuordnen.  Für diesen Zweck wird sowohl im Request,  als auch im Response ein Parameter „businessapp“ angegeben.  Dieser Parameter ist optional und kann mit einem Namen befüllt werden.  Sofern der Aufruf einen Wert enthält, wird dieser auch im Response zurückgegeben.  Somit ist ein nachträgliches Filtern von Aufrufen möglich.  (optional)
     * @param joboe COO-Adresse der zu verwendenden **Organisationseinheit** des angemeldeten Nutzers.\\ \\ Ohne Angabe der joboe und jobposition  wird die eingestellte Standardrolle und OE des Nutzers in der eAkte verwendet.\\ Wenn eine bestimmte jobposition angegeben wird, unter der dieser Aufruf ausgeführt werden soll  muss auch die entsprechende joboe des Nutzers mit angegeben werden.  Das kann z.B. notwendig sein, wenn bestimmte Funktionen nur durch Schriftgutverwalter oder Registratur ausgeführt werden dürfen.\\ Über die OE wird der Mandant zugeordnet und entsprechend die Berechtigungen geprüft.\\ Hinweis:\\ Die COO-Adresse für die joboe muss im Fachverfahren in einer Konfigurationsdatei hintelegt werden, da sie nicht per Schnittstelle ausgelesen werden kann und sich ggf. ändern kann.  Meistens ist die joboe innerhalb eines Fachbereichs für alle Benutzer gleich, so dass sie bei Änderungen einfach zu verwalten ist.  (optional)
     * @param jobposition Referenz der Stelle (**Rolle des Nutzers** in der eAkte). * Schriftgutverwaltung - DocumentManager * Sachbearbeitung - OfficialInCharge * Fachadministration - OpAdm * Registratur - Official * Sekretariat - Secretary * Leitung - Head * Archivar - Registrar * Vorlagenadministrator - TemplateAdministrator * Vorlagenverwaltung - TemplateManager  Es dürfen ausschließlich die o.g. genannten Stellen der LHM verwendet werden.\\ Bitte verwenden Sie die englische Bezeichnung.  Die Rolle Fachadministrator ist bei Schnittstellen nicht erlaubt!  (optional)
     * @return Objekt wurde erfolgreich aktualisiert. (status code 204)
     *         or Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten. (status code 400)
     *         or Falscher oder fehlender technischer Nutzer. (status code 401)
     *         or Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten. (status code 500)
     */
    @Operation(
        operationId = "setSignature",
        summary = "Unterschrift setzen auf Objekten",
        description = "Bei dem Schriftgutobjekt (parameter: objaddress) wird eine definierte Unterschriftenart gesetzt (parameter: signaturetype).  Dabei wird die in der Domäne/Mandanten gültige Unterschriftenregelung beachtet.  Die jeweiligen Vorbedingungen werden vor dem Anbringen der Unterschrift ebenfalls geprüft.  Bitte beachten Sie, dass beim Anbringen der Unterschrift  auch Folge-Aktivitäten ausgeführt werden  (wie z.B. der Bearbeitungsstatus wird korrekt gesetzt).\\ Unterschriften, in denen weitere GUI-Masken angezeigt werden (wie z.B. bei Umschreibungen),  können über diese Soap-Aktion natürlich nicht angebracht werden.  Für entsprechende Aktionen müssen bei Bedarf spezielle Soap-Aktionen entwickelt werden.  Im Kontext des 4-Augen-Löschens sind über diesen Aufruf  nur die dafür freigegebenen Unterschriften „Löschmarkierung setzen“, „Löschmarkierung bestätigen“  und „Löschmarkierung ablehnen“ über die Schnittstelle zulässig,  sofern sie in der gültigen Unterschriftenkonfiguration vorgesehen sind und  alle fachlichen Vorbedingungen erfüllt sind.\\ Die Aktivitäten im Arbeitsvorrat, die nach Löschmarkierung setzen gestartet werden,  werden im Arbeitsvorrat des Technischen Nutzers aufgeräumt,  wenn der AT-Job die bestätigten Löschaufträge aus dem Arbeitsvorrat entfernt. ",
        tags = { "/setSignatureObjects" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Objekt wurde erfolgreich aktualisiert."),
            @ApiResponse(responseCode = "400", description = "Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DmsErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Falscher oder fehlender technischer Nutzer."),
            @ApiResponse(responseCode = "500", description = "Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten.")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = SetSignatureObjectsApi.PATH_SET_SIGNATURE,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> setSignature(
        @NotNull @Parameter(name = "userlogin", description = "**Benutzer-Login** (Benutzer-Kontext für den SOAP-Aufruf)\\ Log-in-Name des Benutzerobjekts ", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "userlogin", required = true) String userlogin,
        @Parameter(name = "objaddress", description = "Objekt-ID (COO-Adresse) des Objekts", required = true, in = ParameterIn.PATH) @PathVariable("objaddress") String objaddress,
        @Parameter(name = "SetSignatureRequestDTO", description = "", required = true) @Valid @RequestBody SetSignatureRequestDTO setSignatureRequestDTO,
        @Parameter(name = "x-anwendung", description = "**Aufrufende Fachanwendung**  Für jeden Aufruf soll ersichtlich sein,  welche Fachanwendung den Aufruf getätigt hat.  Dies soll es ermöglichen, Requests und Responses den entsprechenden Fachverfahren zuzuordnen.  Für diesen Zweck wird sowohl im Request,  als auch im Response ein Parameter „businessapp“ angegeben.  Dieser Parameter ist optional und kann mit einem Namen befüllt werden.  Sofern der Aufruf einen Wert enthält, wird dieser auch im Response zurückgegeben.  Somit ist ein nachträgliches Filtern von Aufrufen möglich. ", in = ParameterIn.HEADER) @RequestHeader(value = "x-anwendung", required = false) Optional<String> xAnwendung,
        @Parameter(name = "joboe", description = "COO-Adresse der zu verwendenden **Organisationseinheit** des angemeldeten Nutzers.\\ \\ Ohne Angabe der joboe und jobposition  wird die eingestellte Standardrolle und OE des Nutzers in der eAkte verwendet.\\ Wenn eine bestimmte jobposition angegeben wird, unter der dieser Aufruf ausgeführt werden soll  muss auch die entsprechende joboe des Nutzers mit angegeben werden.  Das kann z.B. notwendig sein, wenn bestimmte Funktionen nur durch Schriftgutverwalter oder Registratur ausgeführt werden dürfen.\\ Über die OE wird der Mandant zugeordnet und entsprechend die Berechtigungen geprüft.\\ Hinweis:\\ Die COO-Adresse für die joboe muss im Fachverfahren in einer Konfigurationsdatei hintelegt werden, da sie nicht per Schnittstelle ausgelesen werden kann und sich ggf. ändern kann.  Meistens ist die joboe innerhalb eines Fachbereichs für alle Benutzer gleich, so dass sie bei Änderungen einfach zu verwalten ist. ", in = ParameterIn.HEADER) @RequestHeader(value = "joboe", required = false) Optional<String> joboe,
        @Parameter(name = "jobposition", description = "Referenz der Stelle (**Rolle des Nutzers** in der eAkte). * Schriftgutverwaltung - DocumentManager * Sachbearbeitung - OfficialInCharge * Fachadministration - OpAdm * Registratur - Official * Sekretariat - Secretary * Leitung - Head * Archivar - Registrar * Vorlagenadministrator - TemplateAdministrator * Vorlagenverwaltung - TemplateManager  Es dürfen ausschließlich die o.g. genannten Stellen der LHM verwendet werden.\\ Bitte verwenden Sie die englische Bezeichnung.  Die Rolle Fachadministrator ist bei Schnittstellen nicht erlaubt! ", in = ParameterIn.HEADER) @RequestHeader(value = "jobposition", required = false) Optional<String> jobposition,
        @Parameter(hidden = true) final HttpServletRequest servletRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"fehlerQuelle\" : \"INTERN\", \"text\" : \"text\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}